requirements.txt
torch
torchvision
numpy
scikit-learn
tqdm
Pillow
opencv-python
